changes to EmbedderConfig: - move from config.py to new embedder_config.py - add to __init__.py for gamechangerml/configs - update references to this class - add docstrings
