+++ gamechangerml/configs/__init__.py
@@ -1,2 +1,3 @@
 from .s3_config import S3Config
 from .qa_config import QAConfig
+from .embedder_config import EmbedderConfig
