+++ gamechangerml/configs/config.py
@@ -53,20 +53,6 @@ class BertSummConfig:
     }
 
 
-class EmbedderConfig:
-    BASE_MODEL = "msmarco-distilbert-base-v2"
-    MODEL_ARGS = {
-        "min_token_len": 25,
-        "verbose": True,  # for creating LocalCorpus
-        "return_id": True,  # for creating LocalCorpus
-    }
-    FINETUNE = {"shuffle": True, "batch_size": 32, "epochs": 3, "warmup_steps": 100}
-    DEFAULT_THRESHOLD = 0.7  # if no threshold is recommended in evals, this is the default min score for the sent index
-    THRESHOLD_MULTIPLIER = (
-        0.8  # makes the default threshold less strict (to use exact default, set to 1)
-    )
-
-
 class SimilarityConfig:
     BASE_MODEL = "distilbart-mnli-12-3"
